move compiled-object detection code out of distribution_summary
